Use XCDependenciesLoader to delay target dependencies search
